+++ backend/CalculadoraCalorias.Core/Domain/InternalDTO/RefeicaoModeloDTO.cs
@@ -4,7 +4,7 @@ namespace CalculadoraCalorias.Core.Domain.InternalDTO
     {
         public long Id { get; set; }
         public string? Apelido { get; set; }
-        public double? Calorias { get; set; }
+        public int? Calorias { get; set; }
         public double? Proteinas { get; set; }
         public double? Carboidratos { get; set; }
         public double? Gorduras { get; set; }
